+++ tests/test_timestamps.py
@@ -26,8 +26,9 @@ FNS = MappingProxyType({FN: (97373, 87913)})
 
 DEFAULT_CONFIG = {
     "bigger": False,
+    "convert_jpeg_to_jxl": False,
     "convert_to": [],
-    "formats": ["GIF", "JPEG", "PNG", "SVG", "WEBP"],
+    "formats": ["GIF", "JPEG", "JXL", "PNG", "SVG", "WEBP"],
     "keep_metadata": True,
     "ignore": [],
     "near_lossless": False,